vertex-test-app/requirements.txt
flask
gunicorn
google-cloud-aiplatform
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.10
langchain-google-vertexai
faiss-cpu
sentence-transformers
numpy
pandas